should move key event in general setting overlay because setting hud overlay does not need to know all the keys.
