+++ osu.Game.Rulesets.Karaoke/UI/HUD/GeneralSettingOverlay.cs
@@ -1,15 +1,24 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using osu.Framework.Allocation;
+using osu.Framework.Bindables;
+using osu.Framework.Input.Bindings;
+using osu.Framework.Input.Events;
 using osu.Game.Beatmaps;
 using osu.Game.Rulesets.Karaoke.Beatmaps;
+using osu.Game.Rulesets.Karaoke.Configuration;
 using osu.Game.Rulesets.Karaoke.UI.PlayerSettings;
 using osu.Game.Screens.Play.PlayerSettings;
 
 namespace osu.Game.Rulesets.Karaoke.UI.HUD
 {
-    public class GeneralSettingOverlay : RightSideOverlay
+    public class GeneralSettingOverlay : RightSideOverlay, IKeyBindingHandler<KaraokeAction>
     {
+        private readonly BindableInt bindablePitch = new();
+        private readonly BindableInt bindableVocalPitch = new();
+        private readonly BindableInt bindableSaitenPitch = new();
+
         public GeneralSettingOverlay(IBeatmap beatmap)
         {
             // Add common group
@@ -20,6 +29,9 @@ namespace osu.Game.Rulesets.Karaoke.UI.HUD
             // Add translate group if this beatmap has translate
             if (beatmap.AnyTranslate())
                 Add(new TranslateSettings(beatmap.AvailableTranslates()) { Expanded = false });
+
+            // should always present for getting key binding event.
+            AlwaysPresent = true;
         }
 
         public override SettingButton CreateToggleButton() => new()
@@ -29,5 +41,73 @@ namespace osu.Game.Rulesets.Karaoke.UI.HUD
             TooltipText = "Open/Close setting",
             Action = ToggleVisibility
         };
+
+        // should be able to get the key event.
+        protected override bool BlockNonPositionalInput => false;
+
+        // should get key event even it's hide.
+        public override bool PropagateNonPositionalInputSubTree => true;
+
+        public virtual bool OnPressed(KeyBindingPressEvent<KaraokeAction> e)
+        {
+            switch (e.Action)
+            {
+                // Pitch
+                case KaraokeAction.IncreasePitch:
+                    bindablePitch.TriggerIncrease();
+                    break;
+
+                case KaraokeAction.DecreasePitch:
+                    bindablePitch.TriggerDecrease();
+                    break;
+
+                case KaraokeAction.ResetPitch:
+                    bindablePitch.SetDefault();
+                    break;
+
+                // Vocal pitch
+                case KaraokeAction.IncreaseVocalPitch:
+                    bindableVocalPitch.TriggerIncrease();
+                    break;
+
+                case KaraokeAction.DecreaseVocalPitch:
+                    bindableVocalPitch.TriggerDecrease();
+                    break;
+
+                case KaraokeAction.ResetVocalPitch:
+                    bindableVocalPitch.SetDefault();
+                    break;
+
+                // Saiten pitch
+                case KaraokeAction.IncreaseSaitenPitch:
+                    bindableSaitenPitch.TriggerIncrease();
+                    break;
+
+                case KaraokeAction.DecreaseSaitenPitch:
+                    bindableSaitenPitch.TriggerDecrease();
+                    break;
+
+                case KaraokeAction.ResetSaitenPitch:
+                    bindableSaitenPitch.SetDefault();
+                    break;
+
+                default:
+                    return false;
+            }
+
+            return true;
+        }
+
+        public virtual void OnReleased(KeyBindingReleaseEvent<KaraokeAction> e)
+        {
+        }
+
+        [BackgroundDependencyLoader]
+        private void load(KaraokeSessionStatics session)
+        {
+            session.BindWith(KaraokeRulesetSession.Pitch, bindablePitch);
+            session.BindWith(KaraokeRulesetSession.VocalPitch, bindableVocalPitch);
+            session.BindWith(KaraokeRulesetSession.SaitenPitch, bindableSaitenPitch);
+        }
     }
 }
